backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
SQLAlchemy==2.0.37
psycopg2-binary==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
sumy==0.11.0
nltk>=3.9
reportlab==4.0.9
beautifulsoup4==4.12.3
lxml==5.1.0
numpy>=1.24.0
apscheduler==3.10.4

# Generic URL Scraping (Option C - Minimal)
playwright==1.40.0
playwright-stealth==1.0.5
trafilatura==1.6.3
fake-useragent==1.4.0
httpx==0.25.2